Add RoleTemplate mutating webhook
Apply a patch to ensure that the RoleTemplate is encoded when the encoded rules differ from the non-encoded rules. This disparity was causing the following issue:
A RoleTemplate created with an empty list of rules would lead to the creation of a corresponding ClusterRole with a nil list.
When Rancher synchronizes RoleTemplates and ClusterRoles, it detects that an empty list is not equal to a nil list and attempts to update the ClusterRole.